Add support for changing HostKey option
This is to make it possible to change the HostKey location in
sshd_config or to completely leave it out, falling back to the sshd
defaults of using both /etc/ssh/ssh_host_rsa_key and
/etc/ssh/ssh_host_dsa_key.

As the HostKey is hardcoded to the template now it is currently not
possible to connect using for example ssh-dss. This commit changes that
behaviour.